chore: use graph payments payment types instead of uint8

+++ packages/horizon/contracts/payments/GraphPayments.sol
@@ -56,7 +56,7 @@ contract GraphPayments is Multicall, GraphDirectory, GraphPaymentsStorageV1Stora
 
         // Calculate cuts
         uint256 tokensProtocol = tokens.mulPPM(PROTOCOL_PAYMENT_CUT);
-        uint256 delegationFeeCut = _graphStaking().getDelegationFeeCut(receiver, dataService, uint8(paymentType));
+        uint256 delegationFeeCut = _graphStaking().getDelegationFeeCut(receiver, dataService, paymentType);
         uint256 tokensDelegationPool = tokens.mulPPM(delegationFeeCut);
         uint256 totalCut = tokensProtocol + tokensDataService + tokensDelegationPool;
         if (totalCut > tokens) {
